Rename:  더미데이터 이름 수정 (carousel.json -> program-dummy.json)
